Add namespace to types
