+++ packages/vtex-io-logger/src/core/injection/inject-enhanced-logger-to-routes.ts
@@ -1,10 +1,10 @@
 import type { ServiceConfig } from "@vtex/api";
-import type { Logger } from "../../types/logger";
+import type { Ollie } from "../../types/ollie";
 import { enhancedLoggerInjectionMiddlewareFactory } from "./enhanced-logger-injection-middleware-factory";
 
 export function injectEnhancedLoggerToRoutes(
   routes: ServiceConfig<any, any, any>["routes"],
-  { logger }: { logger?: Logger }
+  { logger }: { logger?: Ollie.Logger }
 ) {
   if (!routes) {
     return routes;
