--- packages/vtex-io-logger/src/types/logger.d.ts
@@ -1,8 +1,0 @@
-type LogFn = (message: any) => void;
-
-export interface Logger {
-  debug: LogFn;
-  info: LogFn;
-  warn: LogFn;
-  error: LogFn;
-}
