+++ packages/vtex-io-logger/src/types/ollie.d.ts
@@ -1,0 +1,18 @@
+import type { ParamsContext } from "@vtex/api";
+
+type LogFn = (message: any) => void;
+
+declare namespace Ollie {
+  export interface Logger {
+    debug: LogFn;
+    info: LogFn;
+    warn: LogFn;
+    error: LogFn;
+  }
+
+  interface ParamsContextWithOllie extends ParamsContext {
+    ollie: {
+      logger: Logger;
+    };
+  }
+}
