--- packages/vtex-io-logger/src/types/service.d.ts
@@ -1,8 +1,0 @@
-import type { ParamsContext } from "@vtex/api";
-import type { Logger } from "./logger";
-
-export interface ParamsContextWithOllie extends ParamsContext {
-  ollie: {
-    logger: Logger;
-  };
-}
