+++ packages/vtex-io-logger/tests/core/injection/enhanced-logger-injection-middleware-factory.spec.ts
@@ -1,12 +1,11 @@
 /* eslint-disable @typescript-eslint/no-unsafe-assignment -- necessary for mocks */
 import type { ServiceContext } from "@vtex/api";
 import { enhancedLoggerInjectionMiddlewareFactory } from "../../../src/core/injection/enhanced-logger-injection-middleware-factory";
-import type { Logger } from "../../../src/types/logger";
-import type { ParamsContextWithOllie } from "../../../src/types/service";
+import type { Ollie } from "../../../src/types/ollie";
 
 describe("enhancedLoggerInjectionMiddlewareFactory", () => {
-  let logger: Logger;
-  let ctx: ServiceContext<any, any, ParamsContextWithOllie>;
+  let logger: Ollie.Logger;
+  let ctx: ServiceContext<any, any, Ollie.ParamsContextWithOllie>;
   let next: jest.Mock;
 
   beforeEach(() => {
@@ -28,7 +27,7 @@ describe("enhancedLoggerInjectionMiddlewareFactory", () => {
   });
 
   it("should use ctx.vtex.logger if logger is not provided", async () => {
-    const vtexLogger: Logger = {
+    const vtexLogger: Ollie.Logger = {
       debug: jest.fn(),
       error: jest.fn(),
       info: jest.fn(),
