+++ packages/vtex-io-logger/tests/core/injection/inject-enhanced-logger-to-routes.spec.ts
@@ -1,7 +1,7 @@
 import type { RouteHandler } from "@vtex/api";
 import { enhancedLoggerInjectionMiddlewareFactory } from "../../../src/core/injection/enhanced-logger-injection-middleware-factory";
 import { injectEnhancedLoggerToRoutes } from "../../../src/core/injection/inject-enhanced-logger-to-routes";
-import type { Logger } from "../../../src/types/logger";
+import type { Ollie } from "../../../src/types/ollie";
 
 jest.mock(
   "../../../src/core/injection/enhanced-logger-injection-middleware-factory"
@@ -13,7 +13,7 @@ describe("injectEnhancedLoggerToRoutes", () => {
     error: jest.fn(),
     debug: jest.fn(),
     warn: jest.fn(),
-  } as Logger;
+  } as Ollie.Logger;
 
   const mockLoggerMiddleware = jest.fn();
